+++ src/modules/apps/get-app-environment.ts
@@ -2,6 +2,7 @@ import { isJSON } from "class-validator";
 
 import type { IApp } from "@/entities";
 import type { DeployEnvironment } from "@/interfaces";
+import { formatEnvVars } from "@/plugins/env-var";
 // import { migrateDeployEnvironmentOfSpecificApps } from "@/migration/migrate-app-environment";
 
 /**
@@ -20,11 +21,10 @@ export const getDeployEnvironmentFromJSON = async (app: IApp, env: string) => {
 };
 
 export const getDeployEvironmentByApp = async (app: IApp, env: string) => {
-	// let deployEnvironment = app.deployEnvironment || {};
-	// if (isEmpty(deployEnvironment) || isEmpty(deployEnvironment[env])) {
-	// 	// try to fetch from old CLI version if any...
-	// 	deployEnvironment[env] = await getDeployEnvironmentFromJSON(app, env);
-	// 	// migrateDeployEnvironmentOfSpecificApps({ _id: app._id });
-	// }
-	return ((app.deployEnvironment || {})[env] || {}) as DeployEnvironment;
+	const deployEnvironment = ((app.deployEnvironment || {})[env] || {}) as DeployEnvironment;
+
+	// format environment variables
+	if (deployEnvironment.envVars) deployEnvironment.envVars = formatEnvVars(deployEnvironment.envVars);
+
+	return deployEnvironment;
 };